Iptables management in a separate role
Iptables management extrapolated from the ldap role
and put in a separate role. It is invoked by the role
dependencies option.

Current implementation can take several variable:

ports (this is a range for the destination port), chain and table
(default respectively INPUT and filter).

Nevertheless, the semantic need to be increased in order to cover other situations